requirements.txt
requests
pywin32
python-dotenv
loguru
